+++ backend/canisters/community/impl/src/model/user_event_batch.rs
@@ -1,13 +1,13 @@
 use crate::{can_borrow_state, run_regular_jobs};
 use timer_job_queues::{TimerJobItem, grouped_timer_job_batch};
-use types::{IdempotentEnvelope, UserId};
+use types::{IdempotentEnvelope, Milliseconds, UserId};
 use user_canister::CommunityCanisterEvent;
-use utils::canister::should_retry_failed_c2c_call;
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 grouped_timer_job_batch!(UserEventBatch, UserId, IdempotentEnvelope<CommunityCanisterEvent>, 1000);
 
 impl TimerJobItem for UserEventBatch {
-    async fn process(&self) -> Result<(), bool> {
+    async fn process(&self) -> Result<(), Option<Milliseconds>> {
         if can_borrow_state() {
             run_regular_jobs();
         }
@@ -23,8 +23,8 @@ impl TimerJobItem for UserEventBatch {
         match response {
             Ok(user_canister::c2c_community_canister::Response::Success) => Ok(()),
             Err(error) => {
-                let retry = should_retry_failed_c2c_call(error.reject_code(), error.message());
-                Err(retry)
+                let delay_if_should_retry = delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message());
+                Err(delay_if_should_retry)
             }
         }
     }