+++ backend/canisters/notifications_index/impl/src/model/local_index_event_batch.rs
@@ -1,7 +1,7 @@
 use notifications_index_canister::NotificationsIndexEvent;
 use timer_job_queues::{TimerJobItem, grouped_timer_job_batch};
-use types::{CanisterId, IdempotentEnvelope};
-use utils::canister::should_retry_failed_c2c_call;
+use types::{CanisterId, IdempotentEnvelope, Milliseconds};
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 grouped_timer_job_batch!(
     LocalIndexEventBatch,
@@ -11,7 +11,7 @@ grouped_timer_job_batch!(
 );
 
 impl TimerJobItem for LocalIndexEventBatch {
-    async fn process(&self) -> Result<(), bool> {
+    async fn process(&self) -> Result<(), Option<Milliseconds>> {
         let args = local_user_index_canister::c2c_notifications_index::Args {
             events: self.items.clone(),
         };
@@ -21,8 +21,8 @@ impl TimerJobItem for LocalIndexEventBatch {
         match response {
             Ok(_) => Ok(()),
             Err(error) => {
-                let retry = should_retry_failed_c2c_call(error.reject_code(), error.message());
-                Err(retry)
+                let delay_if_should_retry = delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message());
+                Err(delay_if_should_retry)
             }
         }
     }