+++ backend/canisters/proposals_bot/impl/src/jobs/retrieve_proposals.rs
@@ -9,7 +9,7 @@ use sns_governance_canister::types::ProposalData;
 use std::collections::HashSet;
 use std::time::Duration;
 use types::{C2CError, CanisterId, Proposal};
-use utils::canister::should_retry_failed_c2c_call;
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 pub const NNS_TOPIC_NEURON_MANAGEMENT: i32 = 1;
 pub const NNS_TOPIC_EXCHANGE_RATE: i32 = 2;
@@ -229,7 +229,7 @@ fn handle_proposals_response<R: RawProposal>(governance_canister_id: CanisterId,
                     .nervous_systems
                     .mark_sync_complete(&governance_canister_id, false, now);
 
-                if !should_retry_failed_c2c_call(error.reject_code(), error.message()) {
+                if delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message()).is_none() {
                     state.data.nervous_systems.mark_disabled(&governance_canister_id);
                 }
             });